backend/requirements.txt
fastapi
uvicorn
tensorflow==2.15.0
keras==2.15.0
numpy
pillow
python-multipart